Revert Deposit Product Fix

+++ app/DepositPayment.php
@@ -38,6 +38,7 @@ class DepositPayment extends Model
 
     public function revert($user_id){
         $data = $this->revertData($user_id);
+        
         $this->update([
             'reverted'=>true,
             'reverted_by'=>$user_id
@@ -47,7 +48,7 @@ class DepositPayment extends Model
 
     public function revertData($user_id){
         
-        $data['amount'] = (int) $this->amount;
+        $data['amount'] = (double) $this->amount;
         $data['transaction_number'] = 'RD'.str_replace('.','',microtime(true));
         $data['payment_method_id'] = (int) $this->payment_method_id;
         $data['notes'] = 'Revertion for ' . $this->transaction_number;
